Completed test update, dropped php 7 support
