+++ .github/workflows/tests.yml
@@ -9,7 +9,7 @@ jobs:
 
         strategy:
             matrix:
-                php: ['7.3', '7.4', '8.0', '8.1']
+                php: ['8.0', '8.1']
 
         name: Tests - PHP ${{ matrix.php }}
