+++ composer.json
@@ -17,11 +17,11 @@
         }
     ],
     "require": {
-        "php": "^7.3|^8.0",
+        "php": "^8.0",
         "laravel/nova": "^4.0"
     },
     "require-dev": {
-        "orchestra/testbench": "^6.29"
+        "orchestra/testbench": "^7.0"
     },
     "autoload": {
         "psr-4": {
